+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 💻 Changes & Update
 
 -   Create Slideshow component to show list of recommended komiks
+-   Create KontenX component to show list of komiks horizontaly